Added address information; Added address in db; Added profile.html that realize address creation; Added logs for sql in console; Changed securityFilterChain in SecurityConfig; Added csrf; ------------ Need to add: more tables(crud must have for them),payment system?,liquibase?

+++ src/main/java/com/example/Education/Address.java
@@ -5,8 +5,10 @@ import lombok.Data;
 @Data
 public class Address {
     private Long id_address;
-    private Long id_user;
+    private Integer id_user;
     private String street;
     private String house;
     private String apartment;
+
+
 }
